This change contains a bunch of enhancements to the tests and test framework. The biggest part of it is allowing multithreaded upload of files and validation of transfers. I wont go into details about the test changes though. (#157)
As for the real changes:
I moved the max connections/sessions out to external configuration for both the api and the worker (they are independent settings). In retrospect maybe I could have called them the same thing - we can talk about it during the review.

Changed logback settings for the sshpool logs to go with the changes in the shared library.

Made changes in the SSHDataClient to prevent deadlocks.